python-cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "python-cli"
version = "0.1.0"
description = "Opinionated Python CLI template powered by Typer and uv."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.0",
  "platformdirs>=3.11.0",
  "pyyaml>=6.0.0",
  "tomli>=2.0.1; python_version < \"3.11\"",
  "tomli-w>=1.0.0",
]

[project.scripts]
python-cli = "python_cli.main:app"

[dependency-groups]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "types-PyYAML>=6.0.12",
]
